Unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/codec/trunk@669572 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/codec/binary/Base64.java
@@ -178,7 +178,7 @@ public class Base64 implements BinaryEncoder, BinaryDecoder {
         this.lineLength = lineLength;
         this.lineSeparator = lineSeparator;
         if (lineLength > 0) {
-            this.encodeSize = (byte) (4 + lineSeparator.length);
+            this.encodeSize = 4 + lineSeparator.length;
         } else {
             this.encodeSize = 4;
         }
